requirements.txt
bugsplat==1.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Deprecated==1.2.13
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
linecache2==1.0.0
numpy==1.23.2
pandas==1.4.3
pathlib==1.0.1
Pillow==9.2.0
pip==21.3.1
pycipher==0.5.2
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyperclip==1.8.2
PySimpleGUI==4.60.3
python-dateutil==2.8.2
pytz==2022.2.1
random-address==1.1.1
requests==2.28.1
setuptools==60.2.0
six==1.16.0
smmap==5.0.0
tenacity==5.1.5
tk==0.1.0
traceback2==1.4.0
urllib3==1.26.11
wget==3.2
wheel==0.37.1
wrapt==1.14.1
wxPython==4.2.0
